Fix SFTP sensor moving
